Refactor Frappe MCP Server: Split large file into modules, improve error handling, and enhance documentation
- Split `frappe-api.ts` into smaller modules: `errors.ts`, `api-client.ts`, `auth.ts`, `document-api.ts`, and `schema-api.ts`.
- Implemented improved error handling for authentication and connection errors.
- Added detailed error messages and context for better debugging.
- Created a test script (`test-auth-error.js`) to verify error handling with invalid credentials.
- Enhanced the overall maintainability and readability of the codebase.
